Fixed many compiler errors for ticket #601
git-svn-id: https://svn.code.sf.net/p/numere/source/branches/ticket_601@955 d50cd26f-24d5-421e-9e7d-30b8b4216d91

+++ kernel/core/datamanagement/memory.hpp
@@ -66,7 +66,7 @@ class Memory : public Sorter
 		bool Allocate(size_t _nNCols, bool shrink = false);
 		void createTableHeaders();
 		bool clear();
-		Boundary findValidBoundary(const VectorIndex& _vLine, const VectorIndex& _vCol, long long int i, long long int j) const;
+		Boundary findValidBoundary(const VectorIndex& _vLine, const VectorIndex& _vCol, int i, int j) const;
 		bool retouch1D(const VectorIndex& _vLine, const VectorIndex& _vCol, AppDir Direction);
 		bool retouch2D(const VectorIndex& _vLine, const VectorIndex& _vCol);
 		bool onlyValidValues(const VectorIndex& _vLine, const VectorIndex& _vCol) const;
